Integrated Turrets and Projectiles
Turrets, Projectiles, and Enemies now interact with each other, as is
shown through the program which runs at of the time of this commit.
entity_vars were rewritten to use pointers.
An off-by-one bug, which wrecked havoc on targeting, which wrecked havoc
on spawning, which grinded the tests to a halt, was fixed (health
checks).